# backend/tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swings API",
    "description": "Swings public + admin HTTP API. Mutating endpoints are annotated; GETs are still fully callable but not listed here (see FDN-02 scope).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to deploying host"
    }
  ],
  "paths": {
    "/api/admin/alerts/{id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "operationId": "update_alert",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Alert id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Alert updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistAlert"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "operationId": "delete_alert",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Alert id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alert deleted"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/admin/members/{id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "operationId": "delete_member",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Member id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member deleted"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/members/{id}/billing-portal": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_member_billing_portal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Member id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingPortalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stripe billing portal URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingPortalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Member has no subscription"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/members/{id}/role": {
      "put": {
        "tags": [
          "admin"
        ],
        "operationId": "update_member_role",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Member id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Role updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/members/{id}/subscription/cancel": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_member_subscription_cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Member id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription scheduled to cancel at period end"
          },
          "400": {
            "description": "Member has no subscription"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/members/{id}/subscription/resume": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_member_subscription_resume",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Member id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription cancellation reversed"
          },
          "400": {
            "description": "Member has no subscription"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/admin/watchlists": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "create_watchlist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWatchlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Watchlist created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Watchlist"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/watchlists/{id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "operationId": "update_watchlist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Watchlist id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWatchlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Watchlist updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Watchlist"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "operationId": "delete_watchlist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Watchlist id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Watchlist deleted"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/watchlists/{id}/alerts": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "create_alert",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Watchlist id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Alert created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistAlert"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingPortalRequest": {
        "type": "object",
        "properties": {
          "return_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BillingPortalResponse": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
...
      "CreateWatchlistRequest": {
        "type": "object",
        "required": [
          "title",
          "week_of"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "week_of": {
            "type": "string",
            "format": "date"
          },
          "video_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "published": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
      "RoleUpdate": {
        "type": "object",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        }
      },
...
      "UpdateAlertRequest": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": [
              "string",
              "null"
            ]
          },
          "direction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TradeDirection"
              }
            ]
          },
          "entry_zone": {
            "type": [
              "string",
              "null"
            ]
          },
          "invalidation": {
            "type": [
              "string",
              "null"
            ]
          },
          "profit_zones": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "chart_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateWatchlistRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "week_of": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "video_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "published": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
      "WatchlistAlert": {
        "type": "object",
        "required": [
          "id",
          "watchlist_id",
          "ticker",
          "direction",
          "entry_zone",
          "invalidation",
          "profit_zones",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "watchlist_id": {
            "type": "string",
            "format": "uuid"
          },
          "ticker": {
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/TradeDirection"
          },
          "entry_zone": {
            "type": "string"
          },
          "invalidation": {
            "type": "string"
          },
          "profit_zones": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "chart_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}